Support formatted parameters
This allows statements that cannot be used with PreparedStatement
to instead be formatted manually and passed to JDBC with a plain
Statement. This allows DDL statements to work on databases that
do not support parameterization on the JDBC side
(notably, PostgreSQL DDL).

Affects: https://github.com/io7m/trasco/issues/3